LMNT: renamed test/helpers.h to test/testhelpers.h to avoid ambiguity with helpers.h in src
